Update dbChange event to take handle a change to any db table
